move HomeeNodeEntity to entity.py
