+++ custom_components/homee/__init__.py
@@ -3,8 +3,7 @@
 import logging
 
 from pyHomee import Homee
-from pyHomee.const import AttributeType, NodeProfile
-from pyHomee.model import HomeeAttribute, HomeeNode
+from pyHomee.const import NodeProfile
 import voluptuous as vol
 
 from homeassistant.config_entries import ConfigEntry, ConfigEntryState
@@ -12,16 +11,13 @@ from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant, ServiceCall, callback
 from homeassistant.exceptions import ServiceValidationError
 from homeassistant.helpers import device_registry as dr, entity_registry as er
-from homeassistant.helpers.entity import Entity
 from homeassistant.helpers.typing import ConfigType
 
 from .const import (
     ATTR_ATTRIBUTE,
     ATTR_CONFIG_ENTRY_ID,
-    ATTR_HOMEE_DATA,
     ATTR_NODE,
     ATTR_VALUE,
-    CONF_ADD_HOMEE_DATA,
     CONF_ALL_DEVICES,
     CONF_DOOR_GROUPS,
     CONF_GROUPS,
@@ -31,7 +27,6 @@ from .const import (
     DOMAIN,
     SERVICE_SET_VALUE,
 )
-from .helpers import get_name_for_enum
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -233,152 +228,6 @@ async def async_migrate_entry(hass: HomeAssistant, config_entry: ConfigEntry) ->
         _LOGGER.info("Migration to v%s successful", config_entry.version)
 
     return True
-
-
-class HomeeNodeEntity:
-    """Representation of a Node in Homee."""
-
-    _unrecorded_attributes = frozenset({ATTR_HOMEE_DATA})
-
-    def __init__(
-        self, node: HomeeNode, entry: HomeeConfigEntry
-    ) -> None:
-        """Initialize the wrapper using a HomeeNode and target entity."""
-        self._node = node
-        self._clear_node_listener = None
-        self._attr_unique_id = node.id
-        self._entry = entry
-
-        self._homee_data = {
-            "id": node.id,
-            "name": node.name,
-            "profile": node.profile,
-            "attributes": [{"id": a.id, "type": a.type} for a in node.attributes],
-        }
-
-    async def async_added_to_hass(self) -> None:
-        """Add the homee binary sensor device to home assistant."""
-        self.register_listener()
-
-    async def async_will_remove_from_hass(self):
-        """Cleanup the entity."""
-        self.clear_listener()
-
-    @property
-    def device_info(self):
-        """Holds the available information about the device."""
-        if self.has_attribute(AttributeType.FIRMWARE_REVISION):
-            sw_version = self.attribute(AttributeType.FIRMWARE_REVISION)
-        elif self.has_attribute(AttributeType.SOFTWARE_REVISION):
-            sw_version = self.attribute(AttributeType.SOFTWARE_REVISION)
-        else:
-            sw_version = "undefined"
-
-        return {
-            "identifiers": {
-                # Serial numbers are unique IDs within a specific domain
-                (DOMAIN, self._node.id)
-            },
-            "name": self._node.name,
-            "manufacturer": "unknown",
-            "model": get_name_for_enum(
-                NodeProfile, self._homee_data["profile"]
-            ).lower(),
-            "sw_version": sw_version,
-            "via_device": (DOMAIN, self._entry.unique_id),
-        }
-
-    @property
-    def available(self) -> bool:
-        """Return the availablity of the underlying node."""
-        return self._node.state <= 1
-
-    @property
-    def should_poll(self) -> bool:
-        """Return if the entity should poll."""
-        return False
-
-    @property
-    def raw_data(self):
-        """Return the raw data of the node."""
-        return self._node.raw_data
-
-    @property
-    def extra_state_attributes(self) -> dict[str, dict]:
-        """Return entity specific state attributes."""
-        data = {}
-
-        if self._entry.options.get(CONF_ADD_HOMEE_DATA, False):
-            data[ATTR_HOMEE_DATA] = self._homee_data
-
-        return data if data else None
-
-    async def async_update(self):
-        """Fetch new state data for this node."""
-        homee = self._entry.runtime_data
-        await homee.update_node(self._node.id)
-
-    def register_listener(self):
-        """Register the on_changed listener on the node."""
-        self._clear_node_listener = self._node.add_on_changed_listener(
-            self._on_node_updated
-        )
-
-    def clear_listener(self):
-        """Clear the on_changed listener on the node."""
-        if self._clear_node_listener is not None:
-            self._clear_node_listener()
-
-    def attribute(self, attribute_type):
-        """Try to get the current value of the attribute of the given type."""
-        try:
-            attribute = self._node.get_attribute_by_type(attribute_type)
-        except KeyError:
-            raise AttributeNotFoundException(attribute_type) from None
-
-        # If the unit of the attribute is 'text', it is stored in .data
-        if attribute.unit == "text":
-            return self._node.get_attribute_by_type(attribute_type).data
-
-        return self._node.get_attribute_by_type(attribute_type).current_value
-
-    def get_attribute(self, attribute_type):
-        """Get the attribute object of the given type."""
-        return self._node.get_attribute_by_type(attribute_type)
-
-    def has_attribute(self, attribute_type):
-        """Check if an attribute of the given type exists."""
-        return attribute_type in self._node.attribute_map
-
-    def is_reversed(self, attribute_type) -> bool:
-        """Check if movement direction is reversed."""
-        attribute = self._node.get_attribute_by_type(attribute_type)
-        if hasattr(attribute.options, "reverse_control_ui"):
-            if attribute.options.reverse_control_ui:
-                return True
-
-        return False
-
-    async def async_set_value(self, attribute_type: int, value: float):
-        """Set an attribute value on the homee node."""
-        await self.async_set_value_by_id(self.get_attribute(attribute_type).id, value)
-
-    async def async_set_value_by_id(self, attribute_id: int, value: float):
-        """Set an attribute value on the homee node."""
-        homee = self._entry.runtime_data
-        await homee.set_value(self._node.id, attribute_id, value)
-
-    def _on_node_updated(self, node: HomeeNode, attribute: HomeeAttribute):
-        self.schedule_update_ha_state()
-
-
-class AttributeNotFoundException(Exception):
-    """Raised if a requested attribute does not exist on a homee node."""
-
-    def __init__(self, attributeType) -> None:
-        """Initialize the exception."""
-        self.attributeType = attributeType
-
 
 async def _migrate_old_unique_ids(hass: HomeAssistant, entry_id: str) -> None:
     entity_registry = er.async_get(hass)
