+++ custom_components/homee/lock.py
@@ -8,7 +8,8 @@ from pyHomee.model import HomeeAttribute, HomeeNode
 from homeassistant.components.lock import LockEntity
 from homeassistant.core import HomeAssistant
 
-from . import HomeeConfigEntry, HomeeNodeEntity
+from . import HomeeConfigEntry
+from .entity import HomeeNodeEntity
 from .helpers import get_imported_nodes, get_name_for_enum
 
 _LOGGER = logging.getLogger(__name__)
